embedding-service/requirements.txt
numpy==1.15.4
Flask==1.0.2
flask-restplus==0.12.1
torch==1.1.0
torchvision==0.3.0
pytorch-transformers==1.1.0
gunicorn==19.9.0
pybind11==2.3.0
nmslib==1.8.1
scikit-learn==0.21.3
psycopg2-binary==2.7.6.1
celery==4.3.0
redis==3.3.11
